Rename setmm.h -> setmm_loader.h.
